Update background color of the leaf category detail screen
