+++ app/assets/stylesheets/tablet/tiltedCardComponentStyles.js
@@ -1,15 +1,12 @@
 import {StyleSheet} from 'react-native';
 import color from '../../../themes/color';
-import {screenHorizontalPadding, cardTitleFontSize} from '../../../constants/component_constant';
+import {cardTitleFontSize} from '../../../constants/component_constant';
 import componentUtil from '../../../utils/component_util';
-
-const screenWidth = Dimensions.get('screen').width;
-const cardWidth = ((screenWidth - (screenHorizontalPadding * 2)) / 2) - 8; // 8 is the margin between the card in a row
 
 const tiltedCardComponentStyles = StyleSheet.create({
   container: {
     maxHeight: 188,
-    width: cardWidth,
+    width: componentUtil.getGridCardWidth(),
   },
   tiltedView: {
     backgroundColor: "transparent",
