+++ app/components/leafCategories/LeafCategoryCardListComponent.js
@@ -1,5 +1,5 @@
 import React, {useState} from 'react';
-import {Image, View} from 'react-native';
+import {Image} from 'react-native';
 
 import Category from '../../models/Category';
 import CardWithSoundWaveComponent from '../shared/CardWithSoundWaveComponent';
@@ -8,20 +8,11 @@ const LeafCategoryCardListComponent = (props) => {
   const [playingUuid, setPlayingUuid] = useState(null);
   const categories = Category.getSubCategories(props.category.uuid);
 
-  const renderImage = () => {
-    return (
-      <Image source={props.category.imageSource}
-        resizeMode='contain'
-        style={{width: '100%', height: 200}}
-      />
-    )
-  }
-
   const renderList = () => {
-    return categories.map((item, index) => {
+    return categories.map((category, index) => {
       return <CardWithSoundWaveComponent
                 key={index}
-                item={item}
+                item={category}
                 playingUuid={playingUuid}
                 updatePlayingUuid={(id) => setPlayingUuid(id)}
               />
@@ -30,7 +21,7 @@ const LeafCategoryCardListComponent = (props) => {
 
   return (
     <React.Fragment>
-      {renderImage()}
+      <Image source={props.category.imageSource} resizeMode='contain' style={{width: '100%', height: 200}} />
       {renderList()}
     </React.Fragment>
   )
