+++ app/components/shared/BoldLabelComponent.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Text } from 'react-native';
 
-import {FontFamily, fontFamily} from '../../themes/font';
+import {FontFamily} from '../../themes/font';
 import color from '../../themes/color';
 import {mediumFontSize} from '../../utils/font_size_util';
 
@@ -9,7 +9,7 @@ import {mediumFontSize} from '../../utils/font_size_util';
 // the English text in uppercase only
 const BoldLabelComponent = (props) => {
   return (
-    <Text {...props} style={[{ color: color.lightBlackColor, fontFamily: FontFamily.bold, fontSize: mediumFontSize() }, props.style]}>
+    <Text {...props} style={[{ color: color.lightBlackColor, fontSize: mediumFontSize(), fontFamily: FontFamily.bold }, props.style]}>
       {props.label}
     </Text>
   )
