+++ app/components/shared/CardWithSoundWaveComponent.js
@@ -6,12 +6,12 @@ import CardWithSoundWaveAudioComponent from './cardWithSoundWaves/CardWithSoundW
 import CardWithSoundWaveInfoComponent from './cardWithSoundWaves/CardWithSoundWaveInfoComponent';
 import color from '../../themes/color';
 import {cardElevation, cardBorderRadius} from '../../constants/component_constant';
-import navigationHelper from '../../helpers/navigation_helper';
+import navigationService from '../../services/navigation_service';
 
 const CardWithSoundWaveComponent = (props) => {
   return (
     <Card mode='elevated' elevation={cardElevation} style={styles.container}
-      onPress={() => navigationHelper.navigateCategory(props.item.uuid)}
+      onPress={() => navigationService.navigateCategory(props.item.uuid)}
     >
       <CardWithSoundWaveAudioComponent
         itemUuid={props.item.uuid}
