+++ app/components/shared/cardWithSoundWaves/CardWithSoundWaveAudioComponent.js
@@ -6,8 +6,8 @@ import {useTranslation} from 'react-i18next';
 import AudioWaveButtonComponent from '../AudioWaveButtonComponent';
 import componentUtil from '../../../utils/component_util';
 import audioUtil from '../../../utils/audio_util';
-import translationUtil from '../../../utils/translation_util';
 import {mediumFontSize} from '../../../utils/font_size_util';
+import translationHelper from '../../../helpers/translation_helper';
 
 const CardWithSoundWaveAudioComponent = (props) => {
   const {i18n} = useTranslation();
@@ -47,7 +47,7 @@ const CardWithSoundWaveAudioComponent = (props) => {
         updatePlayingUuid={props.updatePlayingUuid}
         updatePlaySeconds={(seconds) => updatePlaySeconds(seconds)}
       />
-      <Text style={styles.label}>{ translationUtil.translateNumber(displayDuration.toString(), i18n.language) }</Text>
+      <Text style={styles.label}>{ translationHelper.translateNumber(displayDuration.toString(), i18n.language) }</Text>
     </View>
   )
 }
